navigator: split create-app-reactions into per-domain reaction modules
Step 9 of the controllers/ refactor (see docs/refactor-controllers.md):
move the wiring out of one big file and into reactions/{camera,route}.ts.
The split is by primary effect on the map:

- reactions/camera.ts: padding/offset autorun, choose-on-map toggle,
  destinations/routes/manage-stops fit-to-bounds — anything that calls
  fitPoints / setPadding / setOffset / clearPitchAndBearing.
- reactions/route.ts: route-preview render layering and the activeArrow
  step arrow — anything that mutates map sources/layers.

create-app-reactions.ts becomes a thin composer that calls
wireCameraReactions + wireRouteReactions; existing call sites stay
unchanged. ARCHITECTURE.md updated.
